File URL filler (#27)
When an uploaded file is clicked, the name of that file will be placed in the editor-content at the last place the cursor was. So, if someone was to type: img src=" " and click the file they want displayed, it'd place that file's name in the quotes which was the last place of the cursor ( img src="/fileName.jpeg" ), the blob url would then be rendered and the image would be displayed in the preview.